+++ src/Update.h
@@ -5,8 +5,8 @@
 void compute_density(sim_state_t* s, sim_param_t* params);
 void compute_accel(sim_state_t* state, sim_param_t* params);
 
-void leapfrog_step(sim_state_t* s, double dt);
-void leapfrog_start(sim_state_t* s, double dt);
+void leapfrog_step(sim_state_t* s, float dt);
+void leapfrog_start(sim_state_t* s, float dt);
 
 void damp_reflect(int which, float barrier,
                   float* x, float* v, float* vh);
